Block

Description

This creates a type Block and enumerates it to different possible shapes and randomly selects one of them to insert 
at the top of the game view

Documentation

data Block

Block is a data type of List of Integer pairs(coordinates) and a color of the block

validposition

Arguments

:: (Int, Int)
The 'Int,Int' argument is the block position

-> Block
the Block is input Block

-> Bool
the return value is true if it is valid else false

the validposition checks whether the position of the block is within the bounds of the grid or not

blockcolor :: Block -> Color

the blockcolor takes a block as input and assigns a color to it

randomblock :: Double -> Block

the randomblock takes a random seed and returns a block corresponding to it

rotateclock :: Block -> Block

the rotateclock takes a block and rotates it clockwise and returns the rotated block

rotateanti :: Block -> Block

the rotateanti takes a block and rotates it anti clockwise and returns the rotated block

shapeT :: Block

shapeT defines a shape of T

shapeJ :: Block

shapeJ defines a mirror view of the L shape

shapeS :: Block

shapeS defines a mirror view of the z shape

shapeZ :: Block

shapeZ defines a 90degrees rotated Z

shapeA :: Block

shapeA defines a single block of size 1x1

shapeI :: Block

shapeI defines a shape of 1x4 rectangle

shapeL :: Block

shapeL defines a shape of L

shapeO :: Block

shapeO defines a square of size 2x2

blockcheck :: (Int, Int) -> Block -> Bool

the blockcheck takes a coordinate point and a block and checks whether the point lies in the block or not