Gameview

Description

This creates a Grid as a list of rows where each row is a list of atoms and also has functions to clear a filled row and check 
whether the blocks are colliding

Documentation

data Grid

Grid is a List of rows representing the playing space

emptygrid :: Grid

emptygrid is a list of empty rows

atompos :: Grid -> [(Int, Int, Atom)]

atompos assigns the atom coordinates to all the atoms in the grid

blockrender :: Block -> (Int, Int) -> Grid -> Grid

blockrender takes a block and a coordinate point and a grid and returns a grid with the given block in it at the specified location

doescollide :: Block -> (Int, Int) -> Grid -> Bool

doescollide takes a block and a coordinate and a grid and returns true if the given block collides with any atom in the grid

rownumber :: Grid -> [(Int, Row)]

rownumber assigns the row numbers to the rows

atomindex :: Row -> [(Int, Atom)]

atomindex assigns the atom index to the atoms in a single row

data Atom

the Atom is the basic unit of a block which is filled with a particular color

Constructors

Empty

FilledWith Color

atomcol :: Atom -> Color

the atomcol takes an atom and assigns a color to it

clearFilledRows :: Grid -> (Grid, Int)

clearFilledRows takes the grid and returns a grid with the filled rows removed and the number of filled rows removed