Run

Description

This module descirbes the functions to take input and move blocks in the grid accordingly

Documentation

update :: Float -> State -> State

update takes time and a state and updates the state according to the time

eventhandling :: Event -> State -> State

eventhandling takes the input from the keyboard and the current state of the game and moves the blocks accordingly