State

Description

This module encapsulates the state of the grid and defines a initial state of the grid

Documentation

data State

State encapsulates the state of the grid

Constructors

State

Fields

grid :: Grid

time :: Float

dtime :: Float

nextrend :: Float

block :: Block

blockpos :: (Int, Int)

seed :: StdGen

score :: Int

initState :: State

initState takes Nothing and returns the initial state

reset :: State -> State

reset takes a state and returns the initial state with the seed retained